refactor(ai-channel): general restructuring
